python/python-fastapi/simple_pj/requirement.txt
uvicorn==0.19.0
fastapi==0.85.2
SQLAlchemy==1.4.42
bcrypt==4.0.1
python-jose==3.3.0
python-multipart==0.0.5
passlib           1.7.4